Cleanup logging dependencies
removed the global test dependency on slf4j-nop, because it clutters the
dependency information and is in many cases just wrong (e.g. the build-tools
module has no such dependency).

As a consequence, the maven tests now log error messages about missing slf4j
implementations again. For now however, this seems to be the best solution,
since none of the LearnLib modules actually require a logging implementation
as it is merely a dependency for the end-user, who will use LearnLib.

The best example is the 'examples' module, for which a logback runtime
dependency was added.