remove code examples from archetypes
When packaging the archetype, the code examples will be moved to the target
folder of the archetype (because they will be needed when generating a
project from the archetype).

However, all code-analysis plugins will pickup the example code as actual
code of the archetype which we do not want.

Configuring around this is currently not worth the effort. We rather supply
in-tree examples or wiki pages.

This partially reverts fa3f601