EQ Oracle refactoring / clean ups.
Changed the order of type parameters to always start with an Automata when needed (<I, D, A> -> <A, I, D>).
All constructors now have the membership oracle as their first parameter.
Added a bunch of DFA and Mealy versions of the oracles for easier (and type correct) use.
Added the Random Word and Complete Exploration Oracle to the EquivalenceQueries factory.