partially revert 183f0a5
Revert the change to rely on the uploaded Sonatype artifacts of AutomataLib
and use local builds of the development branch again.

The previous change broke building forks of LearnLib that relied on forked versions
of AutomataLib.

The actual problem, that the original commit tried to fix (breaking the coveralls
plugin by clearing the cache to early) should be already be fixed by performing
the upload during the script phase (rather than the after_success phase)