Aggregate membership queries in DT-based learners
Refactors several DT-based learning algorithms (KV, DT [+VPDA variant], TTT [+VPDA variant]) to collect and prepare as many queries as possible before posing them to a membership oracle. The goal of this change is to allow for a better parallelization in case a parallel oracle is used (which can only split a given batch onto different oracle instances).

This also includes some refactorings among the discrimination tree implementations to share as much code as possible.

fixes #55