oracles: allow size estimate for W{,p}Method oracles
Allows to provide an additional parameter for the estimated size of the system under learning, which is used to dynamically adjust the size of the middle part of the W{,p} method test sequence generation throughout the learning process.

Note, this could break existing code, because the second int parameter used in the constructor no longer determines the batch size, but the estimated size.

fixes LearnLib/automatalib#32