partially revert 0f134231, 1f3e66a9 and b9175615
This reverts a majority of the initial implementation of the partial L* algorithm, including the partial observation table and some of the SLI-specific interfaces.

While the concept of state local inputs showed huge improvements regarding query performance, the implementation via specific partial versions of learning algorithms was hard to maintain, as essentially every currently existing learning algorithm would have to be re-implemented.

Instead, this commit introduces the SLI concept as a SUL filter and introduces a special StateLocalInputSULOracle which early-answers queries that would traverse unavailable inputs with a previously specified symbol. This way, queries that would traverse undefined input symbols still won't be executed on the SUL but the SUL appears as a 'total' Mealy system to the learner, allowing one to use every currently existing Mealy learner as-is.

The only 'drawback' of this approach is that the internal datastructures of the learners are bigger (consume more memory) compared to a 'partial' version of the learner, since the system appears to be total when it is not. But since this is often not the limiting factor in practical AAL, this is a worthwhile trade-off.